Move encoder command functions to a different file.
Since we will be using a lot of different encoder commands,
we will keep a separate file to just store all the encoder
command configurations and import them into the main script.
